💡 refactor : eslint 설정, type폴더 생성
