+++ .prettierrc.js
@@ -1,0 +1,10 @@
+module.exports = {
+    singleQuote: false,
+    parser: "typescript",
+    semi: false,
+    printWidth: 120,
+    tabWidth: 2,
+    useTabs: false,
+    arrowParens: "avoid",
+    plugins: ["prettier-plugin-tailwindcss"],
+  }
