--- src/pages/search/search.type.ts
@@ -1,3 +1,0 @@
-export type SearchParams = {
-    searchValue: string;
-};
