+++ src/type/components.type.ts
