+++ src/type/search.type.ts
@@ -1,0 +1,3 @@
+export type SearchParams = {
+    searchValue: string;
+};
